refactor: extracted AuthenticatorData mocks into test files
