+++ AppAttest/Tests/AttestationDecoderTests/AuthenticatorDataTests.swift
@@ -3,22 +3,10 @@ import Foundation
 import Testing
 
 struct AuthenticatorDataTests {
-    var authenticatorData: Data {
-        get throws {
-            try Data(base64Encoded: """
-            CVqj6oy4szHiiDd8cYbSvfsW9hVM3M8PUt8FUQyaY2xAAAAAAGFwcGF0dGVzdGRldmVsb3A
-            AIH1Cj/hcabcKPp9XXIa/WOX0V6E2eg8/GurzszVtNzdSpQECAyYgASFYIECvB8wEheNQFz
-            Zl1SCINwg7rYImcdOd7JXaIXypG14ZIlggqzfw6JMMwuDa1jV6px5GFRJF6DY2PzBKpkHwJ
-            j82/fM=
-            """)
-        }
-    }
-
     @Test("Authenticator Data correctly extracts the relying party hash")
     func rpIDHash() throws {
-        let sut = try AuthenticatorData(
-            rawValue: authenticatorData
-        )
+        let data = try Data(filename: "authenticator-data-valid")
+        let sut = AuthenticatorData(rawValue: data)
         #expect(
             sut.relyingPartyIDHash.base64EncodedString() ==
             "CVqj6oy4szHiiDd8cYbSvfsW9hVM3M8PUt8FUQyaY2w="
@@ -27,20 +15,13 @@ struct AuthenticatorDataTests {
 
     @Test("Authenticator Data correctly extracts the sign count")
     func signCount() throws {
-        let sut = try AuthenticatorData(
-            rawValue: authenticatorData
+        let unused = try AuthenticatorData(rawValue:
+            Data(filename: "authenticator-data-valid")
         )
-        #expect(
-            sut.counter == 0
-        )
+        #expect(unused.counter == 0)
 
-        let largeCount = AuthenticatorData(rawValue:
-            try Data(base64Encoded: """
-            CVqj6oy4szHiiDd8cYbSvfsW9hVM3M8PUt8FUQyaY2xAAAABEGFwcGF0dGVzdGRldmVsb3A
-            AIH1Cj/hcabcKPp9XXIa/WOX0V6E2eg8/GurzszVtNzdSpQECAyYgASFYIECvB8wEheNQFz
-            Zl1SCINwg7rYImcdOd7JXaIXypG14ZIlggqzfw6JMMwuDa1jV6px5GFRJF6DY2PzBKpkHwJ
-            j82/fM=
-            """)
+        let largeCount = try AuthenticatorData(rawValue:
+            Data(filename: "authenticator-data-used")
         )
         #expect(largeCount.counter == 272)
     }
