+++ AppAttest/Tests/AttestationDecoderTests/Resources/authenticator-data-used.txt
@@ -1,0 +1,1 @@
+CVqj6oy4szHiiDd8cYbSvfsW9hVM3M8PUt8FUQyaY2xAAAABEGFwcGF0dGVzdGRldmVsb3AAIH1Cj/hcabcKPp9XXIa/WOX0V6E2eg8/GurzszVtNzdSpQECAyYgASFYIECvB8wEheNQFzZl1SCINwg7rYImcdOd7JXaIXypG14ZIlggqzfw6JMMwuDa1jV6px5GFRJF6DY2PzBKpkHwJj82/fM=
